+++ source/blender/blenlib/intern/math_vec.cc
@@ -144,24 +144,24 @@ mpq3 mpq3::cross_poly(Span<mpq3> poly)
   return n;
 }
 
-uint32_t hash_mpq_class(const mpq_class &value)
+uint64_t hash_mpq_class(const mpq_class &value)
 {
   /* TODO: better/faster implementation of this. */
   return DefaultHash<float>{}(static_cast<float>(value.get_d()));
 }
 
-uint32_t mpq2::hash() const
+uint64_t mpq2::hash() const
 {
-  uint32_t hashx = hash_mpq_class(this->x);
-  uint32_t hashy = hash_mpq_class(this->y);
+  uint64_t hashx = hash_mpq_class(this->x);
+  uint64_t hashy = hash_mpq_class(this->y);
   return hashx ^ (hashy * 33);
 }
 
-uint32_t mpq3::hash() const
+uint64_t mpq3::hash() const
 {
-  uint32_t hashx = hash_mpq_class(this->x);
-  uint32_t hashy = hash_mpq_class(this->y);
-  uint32_t hashz = hash_mpq_class(this->z);
+  uint64_t hashx = hash_mpq_class(this->x);
+  uint64_t hashy = hash_mpq_class(this->y);
+  uint64_t hashz = hash_mpq_class(this->z);
   return hashx ^ (hashy * 33) ^ (hashz * 33 * 37);
 }
